fix: rename TableSkeleton component
