+++ src/app/(protected)/dashboard/loading.tsx
@@ -1,6 +1,6 @@
 // src/app/(protected)/dashboard/loading.tsx
 import { CardSkeleton } from "@/components/CardSkeleton";
-import { TableSkeleton } from "@/components/TablSkeleton";
+import { TableSkeleton } from "@/components/TableSkeleton";
 
 export default function Loading() {
   return (
